🧬 Migrate the base data modal *_TemplatableConfig* into a new single module *template._base* to be more clear the target meaning of object.
