+++ pymock_api/model/api_config/__init__.py
@@ -19,7 +19,8 @@ from .apis import (
 )
 from .base import BaseConfig
 from .item import IteratorItem
-from .template import TemplateConfig, _TemplatableConfig
+from .template import TemplateConfig
+from .template._base import _TemplatableConfig
 from .template._divide import (
     BeDividedableAsTemplatableConfig,
     DivideStrategy,
