+++ test/unit_test/model/api_config/template/_base.py
@@ -1,0 +1,37 @@
+from abc import ABC
+
+from pymock_api.model.api_config import _TemplatableConfig
+
+from ....._values import _Mock_Templatable_Setting
+from .._base import ConfigTestSpec
+
+
+class TemplatableConfigTestSuite(ConfigTestSpec, ABC):
+    def test_apply_template_props_default_value(self, sut: _TemplatableConfig):
+        assert sut.apply_template_props is True
+
+    def test_apply_template_props_should_be_serialize_if_has_in_config(self, sut_with_nothing: _TemplatableConfig):
+        # Given data
+        has_prop_apply_template_props = self._expected_serialize_value().copy()
+        has_prop_apply_template_props.update(_Mock_Templatable_Setting)
+
+        # Run target function
+        deserialized_sut = sut_with_nothing.deserialize(has_prop_apply_template_props)
+        serialized_sut = deserialized_sut.serialize()
+
+        # Verify
+        assert serialized_sut.get("apply_template_props", None) is not None
+        assert serialized_sut["apply_template_props"] is _Mock_Templatable_Setting["apply_template_props"]
+
+    def test_apply_template_props_should_not_be_serialize_if_not_has_in_config(
+        self, sut_with_nothing: _TemplatableConfig
+    ):
+        # Given data
+        not_has_prop_apply_template_props = self._expected_serialize_value().copy()
+
+        # Run target function
+        deserialized_sut = sut_with_nothing.deserialize(not_has_prop_apply_template_props)
+        serialized_sut = deserialized_sut.serialize()
+
+        # Verify
+        assert serialized_sut.get("apply_template_props", None) is None
